8178 add tests for purchase order queries and move purchase order migrations into 2.10 (msupply-foundation/open-msupply#8202)
* Add initial migration fragments

* Update add_purchase_order_tables.rs

* Update add_purchase_order_tables.rs

* Update add_purchase_order_tables.rs

* Update mod.rs

* Add dummy graphql purchase orders

* Update fields, add Status type

* Update add_purchase_order_tables.rs

* Update refresh_dates.rs

* Fix refresh_dates tests

* Reference purchase_order from purchase_order_line

* Add purchase order lines

* Add WIP purchase order row

* Remove submodule reference

* Add item_id to purchase_order_line

* Add line_number field

* Update add_purchase_order_tables.rs

* Add WIP back end purchase orders

* 8105 Purchase order initial schema changes (msupply-foundation/open-msupply#8113)

* Add initial migration fragments
* Update fields, add Status type
* Fix refresh_dates tests
* Reference purchase_order from purchase_order_line
* Remove submodule reference
* Add item_id to purchase_order_line
* Add line_number field

* msupply-foundation/open-msupply#8133 Purchase order front-end boilerplate (msupply-foundation/open-msupply#8134)

* Package boilerplate

* Add basic purchase order queries back end

* Change to item_link_id

* Remove comment

* Mimic other line loader call pattern

* Remove comment

* Fix misnamed database field

* Fix

* Update purchase_order.rs

* Remove comments

* Add purchase order number to type

* Add filtering by store_id

* 8100 add backend purchase order queries (msupply-foundation/open-msupply#8171)

* Add initial migration fragments

* Update add_purchase_order_tables.rs

* Update add_purchase_order_tables.rs

* Update add_purchase_order_tables.rs

* Update mod.rs

* Add dummy graphql purchase orders

* Update fields, add Status type

* Update add_purchase_order_tables.rs

* Update refresh_dates.rs

* Fix refresh_dates tests

* Reference purchase_order from purchase_order_line

* Add purchase order lines

* Add WIP purchase order row

* Remove submodule reference

* Add item_id to purchase_order_line

* Add line_number field

* Update add_purchase_order_tables.rs

* Add WIP back end purchase orders

* Add basic purchase order queries back end

* Change to item_link_id

* Remove comment

* Mimic other line loader call pattern

* Remove comment

* Fix misnamed database field

* Fix

* Update purchase_order.rs

* Remove comments

* Add purchase order number to type

* Add filtering by store_id

---------


* Add additional sort and filter fields

* Fix macros and loader params

* Amend reference to nullable

* Make purchase order number optional

* store_id not nullable and fix repo layer types to match migrations

* Add like filter for supplier

* 8179 add sorting and filtering to purchase order queries (msupply-foundation/open-msupply#8186)

* Add initial migration fragments

* Update add_purchase_order_tables.rs

* Update add_purchase_order_tables.rs

* Update add_purchase_order_tables.rs

* Update mod.rs

* Add dummy graphql purchase orders

* Update fields, add Status type

* Update add_purchase_order_tables.rs

* Update refresh_dates.rs

* Fix refresh_dates tests

* Reference purchase_order from purchase_order_line

* Add purchase order lines

* Add WIP purchase order row

* Remove submodule reference

* Add item_id to purchase_order_line

* Add line_number field

* Update add_purchase_order_tables.rs

* Add WIP back end purchase orders

* Add basic purchase order queries back end

* Change to item_link_id

* Remove comment

* Mimic other line loader call pattern

* Remove comment

* Fix misnamed database field

* Fix

* Update purchase_order.rs

* Remove comments

* Add purchase order number to type

* Add filtering by store_id

* Add additional sort and filter fields

* Fix macros and loader params

* Amend reference to nullable

* Make purchase order number optional

* store_id not nullable and fix repo layer types to match migrations

* Add like filter for supplier

* Change purchase order number to i64 not null

To match msupply-foundation/open-msupply#8196 PR

* Regenerate schema

---------


* Add basic query tests for purchase orders and purchase order lines

* Add tests for purchase order line

* Remove console logs

* msupply-foundation/open-msupply#8101 msupply-foundation/open-msupply#8102 Purchase order list view front-end basics (msupply-foundation/open-msupply#8190)

* Basic query schemas
* Add initial List hook
* Basic list components
* Add item_id to purchase_order_line
* Add line_number field

* 8164 purchase order insert mutation and service (msupply-foundation/open-msupply#8196)

* purchase order insert mutation and service
* update to be supplier_id
* remove name link check

* Change item node to be optional

and fix tests

* msupply-foundation/open-msupply#8165 Create Purchase order UI (msupply-foundation/open-msupply#8212)

* Update schema.ts
* Create purchase order action
* Update usePurchaseOrder.ts
* Disabled row UI
* Fix misnamed parameters

* Remove hardcoded empty string for storeId and make storeId optional for purchase order line loaders (msupply-foundation/open-msupply#8226)

* Test lines upserted but not changelog

* msupply-foundation/open-msupply#8194 Initial Detail View for Purchase Order (msupply-foundation/open-msupply#8254)

* Add detail initial columns
* Update columns.ts
* Add AppBarButtons
* Add Toolbar skeleton
* Footer placeholder
* Fix types
* Add eslint exceptions (temp)
* Remove some commented out
* Move purchaseOrderId to hook
* Remove unused import
* PR review suggestions

* Revert to compulsary item link

* Add purchase order line to changelog table name type

* Revert submodules changes

This reverts commit 00e4cfa5f573a908f4c507d7043d9879a57aa090.

* Use develop cargo

* Move migrations into 2.10

* Add migration fragment imports

* Amend migrations

So as to only break DB once

* Remove lingering item code

---------


Source: https://github.com/msupply-foundation/open-msupply/commit/341d97a4c63aaf26b2495875e86832d35865dc9c

+++ server/graphql/types/src/types/purchase_order_line.rs
@@ -26,9 +26,10 @@ impl PurchaseOrderLineNode {
     pub async fn purchase_order_id(&self) -> &str {
         &self.row().purchase_order_id
     }
-    pub async fn line_number(&self) -> &i32 {
+    pub async fn line_number(&self) -> &i64 {
         &self.row().line_number
     }
+    // TODO should item be optional?
     pub async fn item(&self, ctx: &Context<'_>) -> Result<ItemNode> {
         let loader = ctx.get_loader::<DataLoader<ItemLoader>>();
